add check:hoisted command to theia cli
in order to:
- get rid of usage of ts-node
- allow to check final apps
