+++ package.json
@@ -21,7 +21,6 @@
     "chai": "^4.1.0",
     "chai-string": "^1.4.0",
     "concurrently": "^3.5.0",
-    "cross-env": "^5.2.0",
     "electron-mocha": "~3.5.0",
     "ignore-styles": "^5.0.1",
     "istanbul": "^0.4.5",
@@ -46,9 +45,9 @@
     "webdriverio": "4.9.2"
   },
   "scripts": {
-    "prepare": "yarn prepare:travis && yarn prepare:hoisting && yarn rebuild:clean && yarn build:clean",
+    "prepare": "yarn prepare:travis && yarn rebuild:clean && yarn build:clean && yarn prepare:hoisting",
     "prepare:travis": "node scripts/prepare-travis",
-    "prepare:hoisting": "cross-env THEIA_CHECK_HOISTING_NO_BAIL=true ts-node scripts/check-hoisting.ts",
+    "prepare:hoisting": "theia check:hoisted -s",
     "build": "run build",
     "build:clean": "run prepare",
     "docs": "run docs \"@theia/!(example-)*\"",
